+++ src/Soluble/Japha/Bridge/Exception/NoSuchFieldException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Soluble\Japha\Bridge\Exception;
+
+class NoSuchFieldException extends JavaException
+{
+}